requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
retell-sdk>=4.0.0
google-api-python-client>=2.150.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
pydantic>=2.9.0
pydantic-settings>=2.6.0
python-dotenv>=1.0.0
